+ finalizing insert-delete

+++ frontend/src/app/admin/grid-common/grid-base.component.ts
@@ -2,11 +2,13 @@ import {CellChangedEvent} from 'ag-grid-community/dist/lib/entities/rowNode';
 import {IRepo, PageRequest, Sort} from './i_repo';
 import {GridProperties} from './grid_properties';
 import {InfiniteRowModelModule} from '@ag-grid-community/infinite-row-model';
-import {IGetRowsParams} from 'ag-grid';
 import {Filter, prepareFilter} from "./filter";
-import {InsertGrid, InsertGridProperties} from "./insert-grid";
+import {showErrorMessage, InsertGrid, InsertGridProperties, showDataCommittedMessage} from "./insert-grid";
+import {OnInit} from "@angular/core";
 
-export class GridBaseComponent<T> {
+const PAGE_SIZE = 10
+
+export class GridBaseComponent<T> implements OnInit{
   protected gridApi;
   protected gridColumnApi;
   gridOptions;
@@ -17,9 +19,9 @@ export class GridBaseComponent<T> {
   columnDefs
   rowModelType
   rowSelection
-  deleteButtonEnabled = false
+  isDeleteButtonEnabled = false
+  pageSize
 
-  isInsertGridEnabled = false
   insertGrid: InsertGrid<T> = null
 
   constructor(
@@ -33,10 +35,32 @@ export class GridBaseComponent<T> {
     this.columnDefs = properties.columnDefs
     this.rowModelType = "infinite"
     this.gridOptions = {
-      cacheBlockSize: 10
+      cacheBlockSize: PAGE_SIZE
     }
+    this.pageSize = PAGE_SIZE
     this.rowSelection = 'multiple'
+
+    const insertGridProperties = new InsertGridProperties()
+    let insertGridColDefs = []
+    this.columnDefs
+      .filter((cd) => cd.editable !== false)
+      .forEach((cd)=>{
+        insertGridColDefs.push({
+          headerName: cd.headerName,
+          field: cd.field,
+          valueParser: cd.valueParser
+        })
+      })
+    insertGridProperties.colDefs = insertGridColDefs
+    const onInsertCallback = ()=>{
+      //TODO: чета не работает обновление таблицы
+      this.gridApi.purgeInfiniteCache(null)
+      this.ngOnInit()
+    }
+    this.insertGrid = new InsertGrid<T>(this.repo, onInsertCallback, insertGridProperties)
   }
+
+  ngOnInit(): void {}
 
   onGridReady(params) {
     this.gridApi = params.api;
@@ -47,10 +71,8 @@ export class GridBaseComponent<T> {
 
   datasource() {
     return {
-      getRows: (params: IGetRowsParams) => {
-        console.log('[Datasource] - rows requested by grid: ', params);
+      getRows: (params) => {
         const pageSize = this.gridApi.paginationGetPageSize()
-        console.log(`sort model: ${JSON.stringify(params.sortModel)}`)
         const sorts = params.sortModel.map((s) => {
           return new Sort(s.colId, s.sort.toUpperCase())
         })
@@ -62,8 +84,8 @@ export class GridBaseComponent<T> {
         const filters: Filter[] = Object.keys(params.filterModel)
             .map((column) => prepareFilter(column, params.filterModel[column]))
         const pageRequest = new PageRequest(
-          params.startRow / this.gridOptions.cacheBlockSize,
-          this.gridOptions.cacheBlockSize,
+          params.startRow / pageSize,
+          pageSize,
           sorts)
         this.repo.queryForPage(filters, pageRequest)
           .subscribe({
@@ -73,7 +95,7 @@ export class GridBaseComponent<T> {
                 params.successCallback(data)
               }
               if (errors) {
-                console.log(`errors while data fetching: ${errors}`)
+                showErrorMessage(errors)
                 params.failCallback()
               }
               this.loading = loading
@@ -84,9 +106,6 @@ export class GridBaseComponent<T> {
   }
 
   cellValueChanged(event: CellChangedEvent) {
-    console.trace(`cell update: ${event.oldValue} to ${event.newValue}`)
-    console.trace(event.node.data)
-
     const cleaned = JSON.parse(JSON.stringify(event.node.data))
     delete cleaned.__typename
     this.repo.saveMutation([cleaned])
@@ -94,17 +113,14 @@ export class GridBaseComponent<T> {
         next: ({data, errors}) => {
           console.log(`data updated: ${JSON.stringify(data)}`)
           if (data) {
-            alert("changes committed")
+            showDataCommittedMessage()
           }
           if (errors) {
-            console.log(`errors while data fetching: ${errors}`)
+            showErrorMessage(errors)
           }
         },
         error: err => {
-          alert(`
-                Error occurred while saving changes: ${err}
-                Rollbacking cell editing
-              `);
+          showErrorMessage(err)
           this.rollbackCellChange(event)
         }
       });
@@ -117,55 +133,37 @@ export class GridBaseComponent<T> {
     event.node.setData(data)
   }
 
-  showInsertGrid() {
-    let insertGridColDefs = []
-    this.columnDefs
-      .filter((cd) => cd.editable !== false)
-      .forEach((cd)=>{
-        insertGridColDefs.push({
-          headerName: cd.headerName,
-          field: cd.field,
-          valueParser: cd.valueParser
-        })
-      })
-    const insertGridProperties = new InsertGridProperties()
-    insertGridProperties.colDefs = insertGridColDefs
-    const onInsertCallback = ()=>{
-      //TODO: чета не работает обновление таблицы
-      this.gridApi.purgeInfiniteCache(null)
-    }
-    this.insertGrid = new InsertGrid<T>(this.repo, onInsertCallback, insertGridProperties)
-    this.isInsertGridEnabled = true
-  }
-
   onSelectionChanged($event: any) {
     if(this.gridApi.getSelectedRows().length == 0){
-      this.deleteButtonEnabled = false
+      this.isDeleteButtonEnabled = false
+      return
     }
-    this.deleteButtonEnabled = true
+    this.isDeleteButtonEnabled = true
   }
 
   onDeleteClicked() {
     const selectedRows = this.gridApi.getSelectedRows()
     const idFieldName = this.columnDefs.find((it)=>it.editable === false).field
     const ids = selectedRows.map((it)=> it[idFieldName])
+    if(ids.length == 0){
+      alert("Ничего не выбрано :(")
+      return
+    }
     this.repo.deleteMutation(ids)
       .subscribe({
         next: ({data, errors}) => {
           console.log(`data deleted: ${JSON.stringify(data)}`)
           if (data) {
-            alert("changes committed")
+            showDataCommittedMessage()
+            this.gridApi.deselectAll()
             this.gridApi.purgeInfiniteCache()
           }
           if (errors) {
-            console.log(`errors while data fetching: ${errors}`)
+            showErrorMessage(errors)
           }
         },
         error: err => {
-          alert(`
-                  Error occurred while saving changes: ${err}
-                  Rollbacking cell editing
-                `);
+          showErrorMessage(err)
         }
       });
   }
